♻️  Refactor creature stats, move DescriptionHighlight, prepare for spellStat
